Refactor to resolve circular dependencies
